backend/requirements.txt
fastapi==0.115.2
uvicorn==0.32.0
pymongo==4.10.1
pandas==2.2.3
openpyxl==3.1.5
python-dotenv==1.0.1
spacy==3.7.6
python-Levenshtein==0.26.0
langchain==0.3.4
langchain-huggingface==0.1.0
langchain-community==0.3.3
pydantic==2.9.2
sentence-transformers==3.2.0
huggingface_hub==0.26.1
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.0/en_core_web_sm-3.7.0-py3-none-any.whl